Application structure reorganization
