+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ui.activity.MainActivity">
+    tools:context="com.example.adnrey.giphygifs.ui.activities.MainActivity">
 
     <EditText
         android:id="@+id/queryGifs"
